Added order table and api endpoints.

+++ ThMoCo.Api/DTO/AppUser.cs
@@ -9,7 +9,7 @@ public class AppUser
     public string? UserAuthId { get; set; }
     public string? PhotoUrl { get; set; }
     public string? PhoneNumber { get; set; }
-    public double? Fund { get; set; }
+    public decimal? Fund { get; set; }
     public DateTime? UpdatedAt { get; set; }
     public DateTime? LastLogin { get; set; }
     public string? Provider { get; set; } // e.g., "auth0", "google", "facebook"
